musica de juego y pantalla final. Pantalla final

+++ src/client/renderpiece.h
@@ -12,9 +12,9 @@ class RenderPiece {
     public:
         int x_top;
         int y_top;
-        float prob;
+        int prob;
         bool selected;
-        RenderPiece(SDL2pp::Renderer& renderer, int y, int x, std::string path, float prob);
+        RenderPiece(SDL2pp::Renderer& renderer, int y, int x, std::string path, int prob);
         void copy();
         bool clicked(SDL2pp::Point mousePos);
         void set_position(int x, int y);
